requirements-test.txt
# Testing requirements for SensorThings integration
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-homeassistant-custom-component>=0.12.0
pytest-mock>=3.10.0
aioresponses>=0.7.4
